requirements.txt
google-adk>=0.1.0
google-generativeai>=0.8.0
pydantic>=2.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0